Moved predicate mapping code into its own class.
